Add onViewGoCommand to Home.js controller and reconnect Home.js and List.js controller references.
This commit reconnects the home screen with the rest of the app's
functionality by moving the onViewGoCommand into the new Home.js
controller. (onViewGoCommand sends the user to the next page in the
search process.) This commit also uncomments the home and listview
references in each controller to allow them to refer to one another.