Update API endpoints to be accurate
In the store folder (under Seagrant/app), the location, product, and vendor
files were pulling their data from an incorrect (temporary) source.  This
commit updates where the data is now to be pulled from (from the actual
OSL What's Fresh API).  Tests for the corresponding location, product, and
vendor stores were also updated to reflect the change.

Additionally, there was an error in 'loads multiple datas from an api' in the
product store test file.  'TRUE' was used as a value instead of the correct
lowercase 'true'.  This was fixed.